\name{GammaFDR-package}
\alias{GammaFDR-package}
\alias{GammaFDR}
\docType{package}
\title{
Estimating empirical null distribution and false discovery rate for Gamma statistics
}
\description{
The GammaFDR package provides four methods, maximum likelihhod estimation, mode matching, characteristic function method and smoothed characteristic function method, to estimate the empirical null distribution of Gamma test statistics based on the two group Gamma mixture model. The false discovery rate is estimated based on empirical null parameters and speficied quantitle.
}
\details{
\tabular{ll}{
Package: \tab GammaFDR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-10\cr
License: \tab \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}

\author{
Xing Ren
Maintainer: <xingren@buffalo.edu>
}

\references{

}

\keyword{ empirical null distribution, Gamma mixture, maximum likelihood, mode matching, characteristic function, false discovery rate }
\seealso{
}

\examples{
}
